move audio related func to audio
